Avance en el desarrollo de la interfaz grafica.
AbstractSimplexControlador:
+ Se cambia el parametro dto de solucionarSimplex a tipo String.
+ Se eliminan parametros de entrada dto para evitar dependencias.
AbstractSolucionadorSimplex:
+ Se cambia el tipo de retorno de calcularRadio a String[]
DtoSimplex:
+ Se cambia el tipo de retorno de getOperaciones a String.
Fraccion:
+ Cambios en funciones toString.
PantallaPasoIntermedio:
+ Cambios varios. Revisar diff.

+++ Implementación/modelo/Fraccion.java
@@ -93,7 +93,7 @@ public class Fraccion extends AbstractFraccion {
 
     @Override
     public String toString() {
-        return getNumerador() + "/" + getDenominador();
+        return (int) getNumerador() + "/" + (int) getDenominador();
     }
 
     @Override
@@ -107,7 +107,7 @@ public class Fraccion extends AbstractFraccion {
                && getDenominador() != 1){
            return toString();
        }else{
-           return String.valueOf(getNumerador() / getDenominador());
+           return String.valueOf((int)getNumerador() / (int)getDenominador());
        }
     }
 }
